Fix for mock server and non TLS integration test
Add mock server via mock folder
Add non TLS integration test case

Fixes gouger/stupid#44
Fixes guoger/stupid#46
