+++ azure-pipelines.yml
@@ -48,13 +48,17 @@ jobs:
     strategy:
       matrix:
         FABRIC_20:
-          FABRIC_VERSION: '2.2'
-          INTERGATION_CASE: 'ANDLogic'
-        FABRIC_14:
-          FABRIC_VERSION: '1.4'
+          FABRIC_VERSION: '22'
           INTERGATION_CASE: 'ANDLogic'
         FABRIC_20_ORLogic:
-          FABRIC_VERSION: '2.2'
+          FABRIC_VERSION: '22'
           INTERGATION_CASE: 'ORLogic'
+        FABRIC_14:
+          FABRIC_VERSION: '14'
+          INTERGATION_CASE: 'ANDLogic'
+        MOCK_FABRIC:
+          FABRIC_VERSION: 'mock'
+        LATEST:
+          FABRIC_VERSION: ''
     steps:
       - script: ./test/integration-test.sh $(FABRIC_VERSION) $(INTERGATION_CASE)
