--- cmd/fabric/README.md
@@ -1,3 +1,0 @@
-## Mocked Fabric to test throughput of Stupid
-
-This is a tiny, mocked Fabric that does **nothing**, so it runs really fast. We use this tool to test the limit of `stupid`, and making sure there's no performance regression. Currently, `stupid` could produce ~15k tps.
